DBZ-2834 Define new config item to activate OpenTracing
DBZ-2834 Seperate tracing and non-tracing code, clean ups
