+++ debezium-quarkus-outbox/deployment/src/main/java/io/debezium/outbox/quarkus/deployment/OutboxEventHbmWriter.java
@@ -17,7 +17,7 @@ import org.hibernate.boot.jaxb.hbm.spi.JaxbHbmIdentifierGeneratorDefinitionType;
 import org.hibernate.boot.jaxb.hbm.spi.JaxbHbmRootEntityType;
 import org.hibernate.boot.jaxb.hbm.spi.JaxbHbmSimpleIdType;
 
-import io.debezium.outbox.quarkus.internal.EventDispatcher;
+import io.debezium.outbox.quarkus.internal.DebeziumTracerEventDispatcher;
 import io.debezium.outbox.quarkus.internal.JsonNodeAttributeConverter;
 
 /**
@@ -185,7 +185,7 @@ public class OutboxEventHbmWriter {
 
     private static JaxbHbmBasicAttributeType createTracingSpanAttribute(DebeziumOutboxConfig config) {
         final JaxbHbmBasicAttributeType attribute = new JaxbHbmBasicAttributeType();
-        attribute.setName(EventDispatcher.TRACING_SPAN_CONTEXT);
+        attribute.setName(DebeziumTracerEventDispatcher.TRACING_SPAN_CONTEXT);
         attribute.setNotNull(false);
         attribute.setTypeAttribute("string");
