+++ debezium-quarkus-outbox/runtime/pom.xml
@@ -26,6 +26,7 @@
         <dependency>
             <groupId>io.quarkus</groupId>
             <artifactId>quarkus-smallrye-opentracing</artifactId>
+            <optional>true</optional>
         </dependency>
         <!-- Needed primarily for @Incubating annotation -->
         <dependency>
